Added selection propagation and buttons / layout to media clipboard, fixed clicking/dragging from header component, fixed pianoroll resizing bug, created helper to get all valid extensions, and commented out play/stop function in MainComponent.h.

+++ src/pianoroll/PianoRollComponent.cpp
@@ -118,7 +118,7 @@ void PianoRollComponent::resizeNoteGrid(double lengthInSecs)
 
     if (lengthInSecs > 0)
     {
-        noteGrid.setResolution(getPianoRollContainerWidth() / lengthInSecs);
+        noteGrid.setResolution(static_cast<double>(getPianoRollContainerWidth()) / lengthInSecs);
     }
     else
     {
